Fix #169: Unsuccessful post to BGG
There was a change just two days ago to the XML API documentation dropping the leading "www" on the API request:
https://boardgamegeek.com/wiki/diff/BGG_XML_API2/2342384/2488801

This is in sync with the time that people started coming here to point out that they were getting errors on the app. I have not tested this change because I don't have the environment setup, but I feel like it might fix it.