Use `MutableSet<Int>` for selected items in `SearchResultsActivity`.
This change replaces `SnapshotStateList<SearchResult>` with `MutableSet<Int>` to store the IDs of selected search results. This simplifies the logic for handling selections and improves performance by avoiding unnecessary recompositions.

The `SearchResultListItem` now only allows long click to select an item if it's not already in selection mode.
The `MultiSelectionTopAppBar` now receives the count of selected items instead of the list itself and clears the selection after an action.