Refactor Location.SortType enum to use PascalCase
This commit refactors the `Location.SortType` enum to use PascalCase for its members, changing `NAME` to `Name` and `PLAY_COUNT` to `PlayCount` for consistency with Kotlin coding conventions.

All related usages in `LocationsActivity`, `LocationsViewModel`, and the `applySort` extension function in `Location.kt` have been updated to reflect this change. The default sort order in `LocationsViewModel` is also updated to `Location.SortType.PlayCount`.