This commit introduces a new placeholder migration, `MIGRATION_56_60`, to facilitate the transition for users upgrading from database version 56 to the current Room-based schema.
This migration is necessary because of the error "Fatal Exception: java.lang.IllegalStateException: A migration from 56 to 64 was required but not found. Please provide the necessary Migration path via RoomDatabase.Builder.addMigration(...) or allow for destructive migrations via one of the RoomDatabase.Builder.fallbackToDestructiveMigration* functions.". The new empty migration from version 56 to 60 ensures a smooth database initialization for users on this older version.