Add "Powered by BGG" image to About screen
This commit introduces a "Powered by BGG" image to the "About" screen and updates the splash screen to use the same image.

A new `ImagePreference` custom preference is created to display the image within the preference screen. This preference is then added to `preference_about.xml`. The splash screen's background drawable is also updated to use the new `powered_by_bgg_01_sm.png` image instead of the launcher icon.